+++ tests/app/rdev/framework/http/KernelTest.php
@@ -10,10 +10,12 @@ use RDev\HTTP\Requests\Request;
 use RDev\HTTP\Responses\Response;
 use RDev\HTTP\Responses\ResponseHeaders;
 use RDev\IoC\Container;
-use RDev\Routing\Router;
-use RDev\Routing\Compilers\Compiler;
-use RDev\Routing\Compilers\Parsers\Parser;
 use RDev\Routing\Dispatchers\Dispatcher;
+use RDev\Routing\Router;
+use RDev\Routing\Routes\Compilers\ICompiler;
+use RDev\Routing\Routes\Compilers\Parsers\IParser;
+use RDev\Routing\Routes\CompiledRoute;
+use RDev\Routing\Routes\ParsedRoute;
 use RDev\Tests\Applications\Mocks\MonologHandler;
 use RDev\Tests\HTTP\Middleware\Mocks\HeaderSetter;
 use RDev\Tests\Routing\Mocks\Controller;
@@ -26,7 +28,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testAddingEmptyMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $kernel->addMiddleware([]);
         $this->assertEquals([], $kernel->getMiddleware());
     }
@@ -36,7 +38,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testAddingMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         // Test a single middleware
         $kernel->addMiddleware("foo");
         $this->assertEquals(["foo"], $kernel->getMiddleware());
@@ -50,7 +52,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testDisablingAllMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $kernel->addMiddleware("foo");
         $kernel->disableAllMiddleware();
         $this->assertEquals([], $kernel->getMiddleware());
@@ -61,7 +63,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testDisablingCertainMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $kernel->addMiddleware("foo");
         $kernel->addMiddleware("bar");
         $kernel->onlyDisableMiddleware(["foo"]);
@@ -73,7 +75,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testEnablingCertainMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $kernel->addMiddleware("foo");
         $kernel->addMiddleware("bar");
         $kernel->onlyEnableMiddleware(["foo"]);
@@ -85,7 +87,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testGettingMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $this->assertEquals([], $kernel->getMiddleware());
         $kernel->addMiddleware("foo");
         $this->assertEquals(["foo"], $kernel->getMiddleware());
@@ -96,7 +98,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testHandlingExceptionalRequest()
     {
-        $kernel = $this->getKernel(true);
+        $kernel = $this->getKernel(Request::METHOD_GET, true);
         $request = Request::createFromGlobals();
         $response = $kernel->handle($request);
         $this->assertInstanceOf(Response::class, $response);
@@ -108,7 +110,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testHandlingRequest()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $request = Request::createFromGlobals();
         $response = $kernel->handle($request);
         $this->assertInstanceOf(Response::class, $response);
@@ -120,7 +122,7 @@ class KernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testHandlingWithMiddleware()
     {
-        $kernel = $this->getKernel(false);
+        $kernel = $this->getKernel(Request::METHOD_GET, false);
         $kernel->addMiddleware(HeaderSetter::class);
         $request = Request::createFromGlobals();
         $response = $kernel->handle($request);
@@ -130,21 +132,33 @@ class KernelTest extends \PHPUnit_Framework_TestCase
     /**
      * Gets a kernel to use in testing
      *
+     * @param string $method The HTTP method the routes are valid for
      * @param bool $shouldThrowException True if the router should throw an exception, otherwise false
      * @return Kernel The kernel
      */
-    private function getKernel($shouldThrowException)
+    private function getKernel($method, $shouldThrowException)
     {
         $container = new Container();
-        $routeCompiler = new Compiler(new Parser());
+        $compiledRoute = $this->getMock(CompiledRoute::class, [], [], "", false);
+        $compiledRoute->expects($this->any())->method("isMatch")->willReturn(true);
+        $compiledRoute->expects($this->any())->method("getControllerName")->willReturn(Controller::class);
+        $compiledRoute->expects($this->any())->method("getControllerMethod")->willReturn("noParameters");
+        $compiledRoute->expects($this->any())->method("getMiddleware")->willReturn([]);
+        $compiledRoute->expects($this->any())->method("getPathVariables")->willReturn([]);
+        $parsedRoute = $this->getMock(ParsedRoute::class, [], [], "", false);
+        $parsedRoute->expects($this->any())->method("getMethods")->willReturn([$method]);
+        $routeParser = $this->getMock(IParser::class);
+        $routeParser->expects($this->any())->method("parse")->willReturn($parsedRoute);
+        $routeCompiler = $this->getMock(ICompiler::class);
+        $routeCompiler->expects($this->any())->method("compile")->willReturn($compiledRoute);
 
         if($shouldThrowException)
         {
-            $router = new ExceptionalRouter(new Dispatcher($container), $routeCompiler);
+            $router = new ExceptionalRouter(new Dispatcher($container), $routeCompiler, $routeParser);
         }
         else
         {
-            $router = new Router(new Dispatcher($container), $routeCompiler);
+            $router = new Router(new Dispatcher($container), $routeCompiler, $routeParser);
         }
 
         $router->any("/", Controller::class . "@noParameters");